removed movile testing in the keyboard and added it in a seperate file
